+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Tx.hs
@@ -39,7 +39,7 @@ import Cardano.Ledger.Alonzo.Tx (
   AlonzoEraTx,
   IsPhase2Valid (..),
  )
-import Cardano.Ledger.BaseTypes (StrictMaybe (..), integralToBounded)
+import Cardano.Ledger.BaseTypes (StrictMaybe (..), ToKeyValuePairs (..), integralToBounded)
 import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR (..),
@@ -74,6 +74,8 @@ import Cardano.Ledger.Shelley.Tx (shelleyTxEqRaw)
 import Cardano.Ledger.State
 import Control.DeepSeq (NFData (..), deepseq)
 import Control.Monad.Trans.Fail.String (errorFail)
+import Data.Aeson (FromJSON (..), ToJSON (..), withObject, (.:), (.=))
+import qualified Data.Aeson as Aeson
 import qualified Data.ByteString.Lazy as LBS
 import Data.Int (Int64)
 import Data.List.NonEmpty (NonEmpty)
@@ -329,6 +331,72 @@ dijkstraTxEqRaw tx1 tx2 =
 
 instance EqRaw (Tx l DijkstraEra) where
   eqRaw = dijkstraTxEqRaw
+
+instance
+  ( ToJSON (TxBody TopTx DijkstraEra)
+  , ToJSON (TxWits DijkstraEra)
+  , ToJSON (TxAuxData DijkstraEra)
+  ) =>
+  ToKeyValuePairs (DijkstraTx TopTx DijkstraEra)
+  where
+  toKeyValuePairs DijkstraTx {dtBody, dtWits, dtIsPhase2Valid, dtAuxData} =
+    [ "body" .= dtBody
+    , "wits" .= dtWits
+    , "isPhase2Valid" .= dtIsPhase2Valid
+    , "auxData" .= dtAuxData
+    ]
+
+instance
+  ( ToJSON (TxBody TopTx DijkstraEra)
+  , ToJSON (TxWits DijkstraEra)
+  , ToJSON (TxAuxData DijkstraEra)
+  ) =>
+  ToJSON (Tx TopTx DijkstraEra)
+  where
+  toJSON (MkDijkstraTx tx) = Aeson.object (toKeyValuePairs tx)
+
+instance
+  ( FromJSON (TxBody TopTx DijkstraEra)
+  , FromJSON (TxWits DijkstraEra)
+  , FromJSON (TxAuxData DijkstraEra)
+  ) =>
+  FromJSON (Tx TopTx DijkstraEra)
+  where
+  parseJSON = withObject "Tx TopTx DijkstraEra" $ \o ->
+    fmap MkDijkstraTx $
+      DijkstraTx
+        <$> o .: "body"
+        <*> o .: "wits"
+        <*> o .: "isPhase2Valid"
+        <*> o .: "auxData"
+
+instance
+  ( ToJSON (TxBody SubTx DijkstraEra)
+  , ToJSON (TxWits DijkstraEra)
+  , ToJSON (TxAuxData DijkstraEra)
+  ) =>
+  ToJSON (Tx SubTx DijkstraEra)
+  where
+  toJSON (MkDijkstraTx (DijkstraSubTx {dstBody, dstWits, dstAuxData})) =
+    Aeson.object
+      [ "dstBody" .= dstBody
+      , "dstWits" .= dstWits
+      , "dstAuxData" .= dstAuxData
+      ]
+
+instance
+  ( FromJSON (TxBody SubTx DijkstraEra)
+  , FromJSON (TxWits DijkstraEra)
+  , FromJSON (TxAuxData DijkstraEra)
+  ) =>
+  FromJSON (Tx SubTx DijkstraEra)
+  where
+  parseJSON = withObject "Tx SubTx DijkstraEra" $ \o ->
+    fmap MkDijkstraTx $
+      DijkstraSubTx
+        <$> o .: "dstBody"
+        <*> o .: "dstWits"
+        <*> o .: "dstAuxData"
 
 dijkstraTxL :: Lens' (Tx l DijkstraEra) (DijkstraTx l DijkstraEra)
 dijkstraTxL = lens unDijkstraTx (\x y -> x {unDijkstraTx = y})