+++ eras/conway/impl/CHANGELOG.md
@@ -112,6 +112,7 @@
   - `ConwayTxCert era`
   - `ConwayPlutusPurpose f era`
 * Add `ToJSON` instance for `DefaultVote`.
+* Add `ToJSON` and `FromJSON` instances for `TxBody TopTx ConwayEra`
 
 ### cddl